+++ lib/chains/elementChain.js
@@ -1,4 +1,4 @@
-const {omit, isNil, compose} = require('ramda');
+const {omit, compose} = require('ramda');
 const util = require('util');
 
 // Import utils
@@ -31,14 +31,6 @@ const {
 } = require('../composers');
 const {
 	element,
-	elementExist,
-	elementNotExist,
-	elementVisible,
-	elementMatchProps,
-	elementMatchJS,
-	elementNotMatchJS,
-	elementMatchBRS,
-	elementNotMatchBRS,
 	elementClick,
 	chainRepeat,
 	elementMoveTo,
@@ -47,25 +39,20 @@ const {
 	invalidInputMessage,
 	warnVideoAsElementDeprecation,
 } = require('../texts');
-const {SUBJ_COMPARATOR} = require('../constants/comparator');
 const {validate, validators} = require('../validataion');
 const {
 	processServerResponse,
 	getRequestType,
-	applyTimeout,
 	applyCountAndDelay,
-	applyNegation,
 	applyCountAndDelayAndClicks,
 	applyUntilCondition,
 } = require('../utils/socketChainHelper');
 const {
 	SUBJ_COMPARATOR: SUBJ_COMPARATOR_M,
-	VALUE: VALUE_M,
 } = require('../mappings');
-const {VALUE} = require('../constants/element');
 const logger = require('../utils/logger');
+const {elementWithComparatorToString, elementWithComparatorToJSON} = require('../utils/chainUtils');
 const getOpType = require('../utils/opType').getOpType;
-const {placeholdEmpty} = require('../utils/stringUtils');
 const {video} = require('./videoChain');
 
 /**
@@ -89,27 +76,7 @@ const toString = (data, nameOnly = false) => {
 	let out = element(elName);
 
 	if (data.comparator) {
-		switch (data.comparator.type) {
-			case SUBJ_COMPARATOR.VISIBLE:
-				out = elementVisible(elName);
-				break;
-			case SUBJ_COMPARATOR.EXIST:
-				out = (data.isNegated ? elementNotExist : elementExist)(elName);
-				break;
-			case SUBJ_COMPARATOR.MATCH_JS:
-				out = (data.isNegated ? elementNotMatchJS : elementMatchJS)(elName, data.comparator.val);
-				break;
-			case SUBJ_COMPARATOR.MATCH_BRS:
-				out = (data.isNegated ? elementNotMatchBRS : elementMatchBRS)(elName, data.comparator.val);
-				break;
-			default: {
-				const matches = data.comparator.props.map(
-					one => `  ${one.name} ${one.type} ${placeholdEmpty(VALUE_M[one.val] || one.val)}`
-				);
-
-				out = elementMatchProps(elName, '\n' + matches.join(',\n'));
-			}
-		}
+		out = elementWithComparatorToString(data, elName);
 	} else if (data.isClick) {
 		out = elementClick(elName);
 	} else if (data.isMoveTo) {
@@ -198,42 +165,7 @@ const toJSON = data => {
 
 	// element subject
 	if (data.comparator) {
-		socketMessage.request = applyTimeout({
-			type: 'assert',
-			condition: {subject},
-		}, data);
-
-		socketMessage.request.condition.type = applyNegation(data.comparator.type, data);
-
-		// match js
-		if (data.comparator.type === SUBJ_COMPARATOR.MATCH_JS) {
-			socketMessage.request.condition.val = data.comparator.val;
-		}
-		// match bs
-		if (data.comparator.type === SUBJ_COMPARATOR.MATCH_BRS) {
-			socketMessage.request.condition.val = data.comparator.val;
-		}
-		// match props
-		if (data.comparator.type === SUBJ_COMPARATOR.MATCH) {
-			socketMessage.request.condition.expression = data.comparator.props.map(prop => {
-				const out = {
-					property: prop.name,
-					type: prop.type,
-				};
-
-				if (prop.val === VALUE.REPO) {
-					out.inherited = true;
-				} else {
-					out.val = prop.val;
-				}
-
-				if (!isNil(prop.deviation)) {
-					out.deviation = prop.deviation;
-				}
-
-				return out;
-			});
-		}
+		socketMessage.request = elementWithComparatorToJSON(data, subject);
 	}
 
 	return socketMessage;
