+++ lib/chains/videoChain.js
@@ -1,5 +1,3 @@
-const {isNil} = require('ramda');
-
 // Import utils
 const makeChain = require('../utils/makeChain');
 
@@ -26,30 +24,17 @@ const {
 } = require('../composers');
 const {
 	element,
-	elementExist,
-	elementNotExist,
-	elementVisible,
-	elementMatchProps,
-	elementMatchJS,
-	elementNotMatchJS,
-	elementMatchBRS,
-	elementNotMatchBRS,
 } = require('../texts');
-const {SUBJ_COMPARATOR} = require('../constants/comparator');
 const {
 	processServerResponse,
 	getRequestType,
-	applyTimeout,
-	applyNegation,
 } = require('../utils/socketChainHelper');
 const {
 	SUBJ_COMPARATOR: SUBJ_COMPARATOR_M,
-	VALUE: VALUE_M,
 } = require('../mappings');
-const {VALUE} = require('../constants/element');
 const logger = require('../utils/logger');
+const {elementWithComparatorToJSON, elementWithComparatorToString} = require('../utils/chainUtils');
 const getOpType = require('../utils/opType').getOpType;
-const {placeholdEmpty} = require('../utils/stringUtils');
 
 /**
  * Returns human readable description of the video chain.
@@ -66,27 +51,7 @@ const toString = (data, nameOnly = false) => {
 	let out = element(elName);
 
 	if (data.comparator) {
-		switch (data.comparator.type) {
-			case SUBJ_COMPARATOR.VISIBLE:
-				out = elementVisible(elName);
-				break;
-			case SUBJ_COMPARATOR.EXIST:
-				out = (data.isNegated ? elementNotExist : elementExist)(elName);
-				break;
-			case SUBJ_COMPARATOR.MATCH_JS:
-				out = (data.isNegated ? elementNotMatchJS : elementMatchJS)(elName, data.comparator.val);
-				break;
-			case SUBJ_COMPARATOR.MATCH_BRS:
-				out = (data.isNegated ? elementNotMatchBRS : elementMatchBRS)(elName, data.comparator.val);
-				break;
-			default: {
-				const matches = data.comparator.props.map(
-					one => `  ${one.name} ${one.type} ${placeholdEmpty(VALUE_M[one.val] || one.val)}`
-				);
-
-				out = elementMatchProps(elName, '\n' + matches.join(',\n'));
-			}
-		}
+		out = elementWithComparatorToString(data, elName);
 	}
 
 	return out;
@@ -110,42 +75,7 @@ const toJSON = data => {
 
 	// video subject
 	if (data.comparator) {
-		socketMessage.request = applyTimeout({
-			type: 'assert',
-			condition: {subject},
-		}, data);
-
-		socketMessage.request.condition.type = applyNegation(data.comparator.type, data);
-
-		// match js
-		if (data.comparator.type === SUBJ_COMPARATOR.MATCH_JS) {
-			socketMessage.request.condition.val = data.comparator.val;
-		}
-		// match bs
-		if (data.comparator.type === SUBJ_COMPARATOR.MATCH_BRS) {
-			socketMessage.request.condition.val = data.comparator.val;
-		}
-		// match props
-		if (data.comparator.type === SUBJ_COMPARATOR.MATCH) {
-			socketMessage.request.condition.expression = data.comparator.props.map(prop => {
-				const out = {
-					property: prop.name,
-					type: prop.type,
-				};
-
-				if (prop.val === VALUE.REPO) {
-					out.inherited = true;
-				} else {
-					out.val = prop.val;
-				}
-
-				if (!isNil(prop.deviation)) {
-					out.deviation = prop.deviation;
-				}
-
-				return out;
-			});
-		}
+		socketMessage.request = elementWithComparatorToJSON(data, subject);
 	}
 
 	return socketMessage;
